+++ src/main/kotlin/raytracer/math/Point.kt
@@ -1,12 +1,13 @@
 package raytracer.math
 
-data class Point(val x: Float, val y: Float, val z: Float) {
+data class Point(val x: Double, val y: Double, val z: Double) {
+
     inline operator fun plus(rhs: Vector): Point = Point(x + rhs.x, y + rhs.y, z + rhs.z)
     inline operator fun minus(rhs: Vector): Point = Point(x - rhs.x, y - rhs.y, z - rhs.z)
     inline operator fun minus(rhs: Point): Vector = Vector(x - rhs.x, y - rhs.y, z - rhs.z)
     inline operator fun unaryMinus(): Point = Point(-x, -y, -z)
-    inline operator fun times(scalar: Float): Point  = Point(x * scalar, y * scalar, z * scalar)
-    inline operator fun div(scalar: Float): Point = Point(x / scalar, y / scalar, z / scalar)
+    inline operator fun times(scalar: Double): Point  = Point(x * scalar, y * scalar, z * scalar)
+    inline operator fun div(scalar: Double): Point = Point(x / scalar, y / scalar, z / scalar)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
