+++ src/main/kotlin/raytracer/math/Transform.kt
@@ -15,14 +15,14 @@ data class Transform(val matrix: Matrix, val inverseMatrix: Matrix) {
         val Identity = Transform(Matrix.Identity, Matrix.Identity)
 
         fun translate(v: Vector): Transform = translate(v.x, v.y, v.z)
-        fun translate(dx: Float, dy: Float, dz: Float): Transform =
+        fun translate(dx: Double, dy: Double, dz: Double): Transform =
             Transform(translateMatrix(dx, dy, dz), translateMatrix(-dx, -dy, -dz))
 
         fun scale(sv: Vector): Transform = scale(sv.x, sv.y, sv.z)
-        fun scale(sx: Float, sy: Float, sz: Float): Transform =
+        fun scale(sx: Double, sy: Double, sz: Double): Transform =
             Transform(
                 scaleMatrix(sx, sy, sz),
-                scaleMatrix(1.0f / sx, 1.0f / sy, 1.0f / sz)
+                scaleMatrix(1.0 / sx, 1.0 / sy, 1.0 / sz)
             )
 
         fun rotationX(r: Radians): Transform =
@@ -32,26 +32,26 @@ data class Transform(val matrix: Matrix, val inverseMatrix: Matrix) {
     }
 }
 
-fun translateMatrix(dx: Float, dy: Float, dz: Float): Matrix =
+fun translateMatrix(dx: Double, dy: Double, dz: Double): Matrix =
     Matrix(
-        1.0f, 0.0f, 0.0f, dx,
-        0.0f, 1.0f, 0.0f, dy,
-        0.0f, 0.0f, 1.0f, dz,
-        0.0f, 0.0f, 1.0f, 1.0f
+        1.0, 0.0, 0.0, dx,
+        0.0, 1.0, 0.0, dy,
+        0.0, 0.0, 1.0, dz,
+        0.0, 0.0, 1.0, 1.0
     )
 
 fun rotateXMatrix(r: Radians): Matrix =
     Matrix(
-        1.0f, 0.0f, 0.0f, 0.0f,
-        0.0f, cos(r), -sin(r), 0.0f,
-        0.0f, sin(r), cos(r), 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
+        1.0, 0.0, 0.0, 0.0,
+        0.0, cos(r), -sin(r), 0.0,
+        0.0, sin(r), cos(r), 0.0,
+        0.0, 0.0, 0.0, 1.0
     )
 
-fun scaleMatrix(sx: Float, sy: Float, sz: Float): Matrix =
+fun scaleMatrix(sx: Double, sy: Double, sz: Double): Matrix =
     Matrix(
-        sx, 0.0f, 0.0f, 0.0f,
-        0.0f, sy, 0.0f, 0.0f,
-        0.0f, 0.0f, sz, 0.0f,
-        0.0f, 0.0f, 0.0f, 1.0f
+        sx, 0.0, 0.0, 0.0,
+        0.0, sy, 0.0, 0.0,
+        0.0, 0.0, sz, 0.0,
+        0.0, 0.0, 0.0, 1.0
     )
